backend/generation-service/requirements.txt
# backend/generation-service/requirements.txt
# Dépendances Python pour le microservice Generation Service

# Framework web asynchrone
fastapi~=0.100.0

# Serveur ASGI pour FastAPI
uvicorn[standard]~=0.22.0

# Client HTTP asynchrone pour les appels inter-services (vers KB Service)
httpx~=0.27.0

# Validation des données (dépendance de FastAPI)
pydantic~=2.0.0
pydantic-settings~=2.0.0 # Pour la gestion des paramètres de configuration

# SDKs des Modèles de Langage Avancés (LLMs)
openai~=1.35.0
google-generativeai~=0.5.0
anthropic~=0.27.0

# Dépendances partagées (ne sont pas installées directement via requirements.txt ici,
# mais sont copiées dans l'image Docker et importées comme modules locaux)
# shared (config, models, exceptions, utils)
